refactor(client): tighten types + dedupe sections from review
Simplify-skill review pass on the #120 branch:

- Move ImportMetaEnv augmentation from branding.ts (where it was
  module-scoped and silently ineffective) into a proper ambient
  packages/client/src/vite-env.d.ts so other files reading
  import.meta.env.VITE_* get autocompletion + typecheck.
- Tighten getStatusVariant / getStatusLabel parameter type from string
  to PublicEventStatus (from @c123-live-mini/shared) and remove dead
  default branches and the impossible 'draft' case (draft is excluded
  from the public API by type).
- Drop the misleading "draft and unknown" comment in
  groupEventsByStatus and the narrating JSDoc on getStatusVariant /
  getStatusLabel / EventList.
- Collapse the three near-identical <section><SectionHeader/><EventList/></section>
  blocks in EventListPage into a data-driven STATUS_SECTIONS map.
- Render the empty state directly on the page instead of leaking into
  EventList by passing events={[]}.
- Add width/height attributes to the event thumbnail <img> for CLS
  stability and drop loading="lazy" since the thumbnails sit
  above-the-fold on the homepage.

Refs #120
